% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphfunctions.R
\name{graph_boi}
\alias{graph_boi}
\title{Raincloud plot}
\usage{
graph_boi(x, y)
}
\arguments{
\item{x}{numerical vector}

\item{y}{character or factor vector on how to split the data}
}
\value{
Plot of x vs y
}
\description{
Raincloud plot
}
\examples{
x <- c(1, 2, 3, 4)
y <- c("a", "a", "b", "b")
}
